Move Position back to model
