+++ nautilus_core/accounting/src/account/base.rs
@@ -21,6 +21,7 @@ use nautilus_model::events::account::state::AccountState;
 use nautilus_model::events::order::filled::OrderFilled;
 use nautilus_model::identifiers::account_id::AccountId;
 use nautilus_model::instruments::Instrument;
+use nautilus_model::position::Position;
 use nautilus_model::types::balance::AccountBalance;
 use nautilus_model::types::currency::Currency;
 use nautilus_model::types::money::Money;
@@ -28,8 +29,6 @@ use nautilus_model::types::price::Price;
 use nautilus_model::types::quantity::Quantity;
 use pyo3::prelude::*;
 use rust_decimal::prelude::ToPrimitive;
-
-use crate::position::Position;
 
 #[derive(Debug)]
 #[cfg_attr(
