+++ nautilus_core/accounting/src/lib.rs
@@ -14,7 +14,6 @@
 // -------------------------------------------------------------------------------------------------
 
 pub mod account;
-pub mod position;
 #[cfg(test)]
 pub mod stubs;
