+++ nautilus_core/model/src/lib.rs
@@ -22,6 +22,7 @@ pub mod instruments;
 pub mod macros;
 pub mod orderbook;
 pub mod orders;
+pub mod position;
 pub mod types;
 
 #[cfg(feature = "ffi")]
